+++ torch/nn/qat/__init__.py
@@ -1,1 +1,7 @@
+# flake8: noqa: F401
+r"""QAT Modules
+
+This package is in the process of being deprecated.
+Please, use `torch.ao.nn.qat` instead.
+"""
 from .modules import *  # noqa: F403
