+++ torch/nn/qat/dynamic/__init__.py
@@ -1,1 +1,7 @@
+# flake8: noqa: F401
+r"""QAT Dynamic Modules
+
+This package is in the process of being deprecated.
+Please, use `torch.ao.nn.qat.dynamic` instead.
+"""
 from .modules import *  # noqa: F403
