+++ torch/nn/qat/dynamic/modules/__init__.py
@@ -1,3 +1,9 @@
+# flake8: noqa: F401
+r"""QAT Dynamic Modules
+
+This package is in the process of being deprecated.
+Please, use `torch.ao.nn.qat.dynamic.modules` instead.
+"""
 from .linear import Linear
 
 __all__ = ["Linear"]
